Use different store for dev and production environment
